requirements.txt
annotated-types==0.7.0
anyio==4.9.0
ariadne==0.26.1
bcrypt==4.3.0
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
graphql-core==3.2.5
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
pyasn1==0.4.8
pydantic==2.11.2
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
